requirements.txt
aionotify==0.2.0
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
kombu==4.5.0
pytz==2018.9
PyYAML==5.1
redis==3.2.1
vine==1.3.0
